requirements.txt
python-telegram-bot==13.0
requests==2.28.1
python-dotenv==0.21.0
argparse==1.4.0